+++ persistent-postgresql/ChangeLog.md
@@ -1,5 +1,13 @@
 # Changelog for persistent-postgresql
 
+## 2.13.4.0 (unreleased)
+
+* [#1341](https://github.com/yesodweb/persistent/pull/1341)
+    * Add `SqlBackendHooks` to allow for instrumentation of queries.
+
+* [#1327](https://github.com/yesodweb/persistent/pull/1327)
+    * Update backend to support new `StatementCache` interface
+
 ## 2.13.3.0
 
 * [#1349](https://github.com/yesodweb/persistent/pull/1349)